+++ docs/integrations/astro-mermaid.ts
@@ -1,0 +1,27 @@
+// yoinked from https://github.com/paularmstrong/onerepo
+
+import type { RemarkPlugin } from "@astrojs/markdown-remark";
+import { visit } from "unist-util-visit";
+
+const escapeMap: Record<string, string> = {
+  "&": "&amp;",
+  "<": "&lt;",
+  ">": "&gt;",
+  '"': "&quot;",
+  "'": "&#39;",
+};
+
+const escapeHtml = (str: string) =>
+  str.replace(/[&<>"']/g, (c) => escapeMap[c] ?? c);
+
+export const mermaid: RemarkPlugin<[]> = () => (tree) => {
+  visit(tree, "code", (node) => {
+    if (node.lang !== "mermaid") return;
+
+    // @ts-ignore
+    node.type = "html";
+    node.value = `<div class="mermaid" data-content="${escapeHtml(
+      node.value,
+    )}"><p>Loading graph...</p></div>`;
+  });
+};