+++ fiberplane-models/src/pagerduty.rs
@@ -29,6 +29,7 @@ use fp_bindgen::prelude::Serializable;
 #[cfg(feature = "axum_06")]
 use http_02::StatusCode;
 
+use crate::errors::GeneralError;
 #[cfg(feature = "axum_07")]
 use http_1::StatusCode;
 
@@ -359,58 +360,7 @@ impl axum_07::response::IntoResponse for PagerDutyReceiverDeleteError {
 }
 
 /// Errors that can occur when listing PagerDuty receivers.
-#[derive(Debug, Deserialize, Serialize, Clone, Copy, PartialEq, Eq, Error)]
-#[cfg_attr(
-    feature = "fp-bindgen",
-    derive(Serializable),
-    fp(rust_module = "fiberplane_models::pagerduty")
-)]
-#[non_exhaustive]
-#[serde(tag = "error", content = "details", rename_all = "snake_case")]
-pub enum PagerDutyReceiverListError {
-    #[error("Unknown error occurred")]
-    InternalServerError,
-
-    /// Common auth errors.
-    #[error(transparent)]
-    Auth(AuthError),
-}
-
-impl PagerDutyReceiverListError {
-    #[cfg(any(feature = "axum_06", feature = "axum_07"))]
-    fn status_code(&self) -> StatusCode {
-        match self {
-            PagerDutyReceiverListError::InternalServerError => StatusCode::INTERNAL_SERVER_ERROR,
-            PagerDutyReceiverListError::Auth(auth_err) => auth_err.status_code(),
-        }
-    }
-}
-
-impl From<AuthError> for PagerDutyReceiverListError {
-    fn from(value: AuthError) -> Self {
-        PagerDutyReceiverListError::Auth(value)
-    }
-}
-
-#[cfg(feature = "axum_06")]
-impl axum_06::response::IntoResponse for PagerDutyReceiverListError {
-    fn into_response(self) -> axum_06::response::Response {
-        let body = serde_json::to_string(&self).expect("unable to serialize error body");
-        let status_code = self.status_code();
-
-        (status_code, body).into_response()
-    }
-}
-
-#[cfg(feature = "axum_07")]
-impl axum_07::response::IntoResponse for PagerDutyReceiverListError {
-    fn into_response(self) -> axum_07::response::Response {
-        let body = serde_json::to_string(&self).expect("unable to serialize error body");
-        let status_code = self.status_code();
-
-        (status_code, body).into_response()
-    }
-}
+pub type PagerDutyReceiverListError = GeneralError;
 
 #[derive(Debug, Deserialize, Serialize, Clone, Copy, PartialEq, Eq, IntoStaticStr)]
 #[cfg_attr(feature = "clap", derive(clap::ValueEnum))]