+++ CHANGELOG.md
@@ -15,11 +15,13 @@ major version hasn't changed.
 
 ## unreleased
 
+- `fiberplane-charts`: Add configuration to MetricsChart for number of ticks you'd prefer to see on an axis (defaults to previously hard-coded values)
 - `fiberplane-models`: Add extra types to front matter values, and add a value validation method to
   front matter schema entries.
 - `fiberplane-models`: The `updated_at` and `created_at` fields in the `IntegrationSummary` struct are now optional (#171)
 - Add `PagerDutyWebhooks` to `WorkspaceIntegrationId` (#189)
 - Replace a secret provided by Fiberplane with one provided by PagerDuty (#196)
+- Add models required for GitHub app workspace-level integration (#198)
 
 ## [v1.0.0-beta.14] - 2024-03-07